reuse semVer constraint
