+++ grype/version/semantic_constraint.go
@@ -14,6 +14,7 @@ var normalizer = strings.NewReplacer(".alpha", "-alpha", ".beta", "-beta", ".rc"
 type semanticConstraint struct {
 	raw        string
 	constraint hashiVer.Constraints
+	checker    func(hashiVer.Constraints, *Version) (bool, error)
 }
 
 func newSemanticConstraint(constStr string) (semanticConstraint, error) {
@@ -31,11 +32,19 @@ func newSemanticConstraint(constStr string) (semanticConstraint, error) {
 	return semanticConstraint{
 		raw:        normalized,
 		constraint: constraints,
+		checker:    semVerifier,
 	}, nil
 }
 
-func (c semanticConstraint) supported(format Format) bool {
-	return format == SemanticFormat
+func semVerifier(constraint hashiVer.Constraints, version *Version) (bool, error) {
+	if SemanticFormat != version.Format {
+		return false, fmt.Errorf("(semantic) unsupported format: %s", version.Format)
+	}
+
+	if version.rich.semVer == nil {
+		return false, fmt.Errorf("no rich semantic version given: %+v", version)
+	}
+	return constraint.Check(version.rich.semVer.verObj), nil
 }
 
 func (c semanticConstraint) Satisfied(version *Version) (bool, error) {
@@ -50,14 +59,11 @@ func (c semanticConstraint) Satisfied(version *Version) (bool, error) {
 		return true, nil
 	}
 
-	if !c.supported(version.Format) {
-		return false, fmt.Errorf("(semantic) unsupported format: %s", version.Format)
+	if c.checker == nil {
+		return false, fmt.Errorf("check function is undefined")
 	}
 
-	if version.rich.semVer == nil {
-		return false, fmt.Errorf("no rich semantic version given: %+v", version)
-	}
-	return c.constraint.Check(version.rich.semVer.verObj), nil
+	return c.checker(c.constraint, version)
 }
 
 func (c semanticConstraint) String() string {
